# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SportReserve",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-sport-reserve.juvhost.com"
        }
    ],
    "paths": {
...
        "/api/fields": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Get list of fields",
                "description": "Returns list of fields",
                "operationId": "getFieldsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Field"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fields"
                ],
                "summary": "Store a new field",
                "description": "Stores a new field and returns the field data",
                "operationId": "storeField",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/fields/{id}": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Get field information",
                "description": "Returns field data",
                "operationId": "getFieldById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Field not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fields"
                ],
                "summary": "Delete an existing field",
                "description": "Deletes an existing field and returns success message",
                "operationId": "deleteField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Fields"
                ],
                "summary": "Update an existing field",
                "description": "Updates an existing field and returns the updated data",
                "operationId": "updateField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFieldRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Field"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreFieldRequest": {
                "title": "Store Field Request",
                "required": [
                    "name",
                    "location",
                    "type",
                    "hourly_rate"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the field",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the field",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type of the field",
                        "type": "string"
                    },
                    "hourly_rate": {
                        "description": "Hourly rate for renting the field",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "UpdateFieldRequest": {
                "title": "Update Field Request",
                "properties": {
                    "name": {
                        "description": "Name of the field",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the field",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type of the field",
                        "type": "string"
                    },
                    "hourly_rate": {
                        "description": "Hourly rate for renting the field",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "Field": {
                "title": "Field",
                "required": [
                    "name",
                    "location",
                    "type",
                    "hourly_rate"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the field",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the field",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the field",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type of the field",
                        "type": "string"
                    },
                    "hourly_rate": {
                        "description": "Hourly rate for renting the field",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}